Fix: Replaced usage of `[0..<3]` on `simd_float4` with `xyz: simd_float3` extension
Wasn't compiling with Swift 5.7 compiler in 4.2 mode, because apparently the `func subscript(Range<Int>) -> ArraySlice<Element>` was removed from `simd_float4` at some point.  _(This isn't unheard of— Swift's SIMD library has had a bit of a rocky history of renames and changes that didn't respect backwards-compat that most of the rest of Swift obeys.  Personally, I got fed up of it a few years ago and wrote my own Swift wrapper lib for the C simd functions, so I feel the pain.)_

Replaced the code doing an range-subscript array slice, creating a new array, and initing a new `float3` with a simple extension method on `simd_float4` that provides a shader-like `var xyz: simd_float3` property, and using that instead.  Should work for all versions of Swift 4.x or 5.x.